+++ common/rmi/src/main/java/is/codion/common/rmi/server/DefaultServerConfiguration.java
@@ -108,7 +108,7 @@ final class DefaultServerConfiguration implements ServerConfiguration {
 	}
 
 	@Override
-	public Collection<String> auxiliaryServerFactory() {
+	public Collection<String> auxiliaryServerFactories() {
 		return auxiliaryServerFactories;
 	}