+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 ### is.codion.common.rmi
 - AuxiliaryServerFactory.createServer() renamed create().
 - SerializationFilterFactory.createObjectInputFilter() renamed create().
+- ServerConfiguration.auxiliaryServerFactory() renamed auxiliaryServerFactories().
 ### is.codion.swing.common.ui
 - MenuItemBuilder, unused type parameter removed.
 - Control.copy() return type fixed.